backend/requirements.txt
paste
pyramid
pyramid_chameleon
pyramid_debugtoolbar
pyramid_tm
pyramid_mako
SQLAlchemy
transaction
zope.sqlalchemy
waitress
psycopg2
python-gcm
git+git://github.com/mdellavo/lockbox.git@master
WebTest
pytest
pytest-cov
mock